+++ DataExplorer/src/gde/utils/GraphicsUtils.java
@@ -102,15 +102,15 @@ public class GraphicsUtils {
 			dist = dist * -1;
 		}
 		
-		double deltaMainTickValue = deltaScaleValue / numberTicks; //deltaScale / numberTicks;
+		double deltaMainTickValue = deltaScaleValue / (double)numberTicks; //deltaScale / numberTicks;
 		log.log(Level.FINE, "minScaleValue = " + minScaleValue + "; maxScaleValue = " + maxScaleValue + ", deltaMainTickValue = " + deltaMainTickValue);
-		double deltaMainTickPixel = deltaScaleValue * height / deltaScale / numberTicks; //1.0 * height / numberTicks;
-		double deltaPosMini = deltaMainTickPixel / miniticks;
+		double deltaMainTickPixel = deltaScaleValue * (double)height / deltaScale / (double)numberTicks; //1.0 * height / numberTicks;
+		double deltaPosMini = deltaMainTickPixel / (double)miniticks;
 
 		//draw mini ticks below first main tick
-		int yTickPositionMin = Double.valueOf(y0 - (Math.abs(minScaleValue - minValue) * height / deltaScale)).intValue(); //new Double(y0 - i * deltaMainTickPixel).intValue();
+		double yTickPositionMin = (double)y0 - (Math.abs(minScaleValue - minValue) * ((double)height / deltaScale)); //new Double(y0 - i * deltaMainTickPixel).intValue();
 		for (int j = 1; j < miniticks; j++) {
-			int yPosMini = yTickPositionMin + (int)(j * deltaPosMini);
+			int yPosMini = (int) (yTickPositionMin + ((double)j * deltaPosMini));
 			if (yPosMini >= y0) break;
 			log.log(Level.FINEST, "yTickPosition=" + yTickPositionMin + ", xPosMini=" + yPosMini); //$NON-NLS-1$ //$NON-NLS-2$
 			gc.drawLine(x0, yPosMini, x0 - ticklength / 2, yPosMini);
@@ -118,7 +118,7 @@ public class GraphicsUtils {
 		// draw main ticks and mini ticks
 		for (int i=0; i <= numberTicks; i++) {
 			//draw the main scale, length = 5 and gap to scale = 2
-			int yTickPosition = Double.valueOf(yTickPositionMin - i * deltaMainTickPixel).intValue();
+			int yTickPosition = (int) (yTickPositionMin - (double)i * deltaMainTickPixel);
 			gc.drawLine(x0, yTickPosition, x0 - ticklength, yTickPosition);
 			if (isBuildGridVector) horizontalGrid.add(yTickPosition);
 			//draw the sub scale according number of miniTicks
@@ -131,9 +131,9 @@ public class GraphicsUtils {
 			drawTextCentered(df.format(minScaleValue + i * deltaMainTickValue), x0 - ticklength - gap - dist, yTickPosition, gc, SWT.HORIZONTAL);
 		}
 		//draw mini ticks above first main tick
-		int yTickPositionMax = Double.valueOf(yTickPositionMin - numberTicks * deltaMainTickPixel).intValue();
-		for (int j = 1; j < miniticks; j++) {
-			int yPosMini = yTickPositionMax - (int)(j * deltaPosMini);
+		double yTickPositionMax = yTickPositionMin - (double)numberTicks * deltaMainTickPixel;
+		for (double j = 1; j < miniticks; j++) {
+			int yPosMini = (int) (yTickPositionMax - (j * deltaPosMini));
 			if (yPosMini < yTop-1) break;
 			log.log(Level.FINEST, "yTickPosition=" + yTickPositionMax + ", xPosMini=" + yPosMini); //$NON-NLS-1$ //$NON-NLS-2$
 			gc.drawLine(x0, yPosMini, x0 - ticklength / 2, yPosMini);
